python: moved ecl.ecl.rft to ecl.rft.
moved ecl_rft and ecl_rft_cell to ecl.rft.

ecl_rft and ecl_rft_cell tied to ecl.rft.

...
