+++ python/ecl/CMakeLists.txt
@@ -6,12 +6,13 @@ set(PYTHON_SOURCES
 add_python_package("python.ecl"  ${PYTHON_INSTALL_PREFIX}/ecl "${PYTHON_SOURCES}" True)
 
 add_subdirectory(eclfile)
+add_subdirectory(grid)
 add_subdirectory(faults)
+add_subdirectory(rft)
 #add_subdirectory(ecl)
 add_subdirectory(geo)
 add_subdirectory(test)
 add_subdirectory(util)
-add_subdirectory(grid)
 #add_subdirectory(well)
 
 configure_file(ecl_lib_info_build.py.in   ${PROJECT_BINARY_DIR}/${PYTHON_INSTALL_PREFIX}/ecl/__ecl_lib_info.py )
