--- python/ecl/ecl/rft/__init__.py
@@ -1,1 +1,0 @@
-from .well_trajectory import WellTrajectory
