+++ python/ecl/rft/__init__.py
@@ -1,0 +1,4 @@
+from .well_trajectory import WellTrajectory
+from .ecl_rft_cell import EclPLTCell, EclRFTCell
+from .ecl_rft import EclRFT, EclRFTFile
+
