+++ python/ecl/rft/ecl_rft.py
@@ -24,7 +24,7 @@ from cwrap import BaseCClass
 from ecl import EclPrototype
 from ecl.util import monkey_the_camel
 from ecl.util import CTime
-from ecl.ecl import EclRFTCell, EclPLTCell
+from ecl.rft import EclRFTCell, EclPLTCell
 
 class EclRFT(BaseCClass):
     """The EclRFT class contains the information for *one* RFT.
